added 1.8 support and dropped 1.6 support
