+++ .coveragerc
@@ -3,7 +3,6 @@ branch = True
 omit =
     */migrations/*
     entity/version.py
-    entity/apps.py
 source = entity
 [report]
 exclude_lines =
