+++ entity/__init__.py
@@ -1,6 +1,8 @@
 # flake8: noqa
 from .config import EntityConfig, entity_registry, register_entity
 from .version import __version__
-from .models import Entity, EntityKind, EntityRelationship, turn_on_syncing, turn_off_syncing, sync_entities
+from .models import Entity, EntityKind, EntityRelationship
+from .signal_handlers import turn_on_syncing, turn_off_syncing
+from .sync import sync_entities
 
-django_app_config = 'entity.apps.EntityConfig'
+default_app_config = 'entity.apps.EntityConfig'
