+++ entity/apps.py
@@ -3,4 +3,8 @@ from django.apps import AppConfig
 
 class EntityConfig(AppConfig):
     name = 'entity'
-    verbose_name = "Django Entity"
+    verbose_name = 'Django Entity'
+
+    def ready(self):
+        import entity.signal_handlers
+        assert(entity)
