+++ run_tests.py
@@ -10,8 +10,7 @@ from settings import configure_settings
 
 # Configure the default settings and setup django
 configure_settings()
-if django.VERSION[1] >= 7:
-    django.setup()
+django.setup()
 
 # Django nose must be imported here since it depends on the settings being configured
 from django_nose import NoseTestSuiteRunner
@@ -32,7 +31,6 @@ def run_tests(*test_args, **kwargs):
 if __name__ == '__main__':
     parser = OptionParser()
     parser.add_option('--verbosity', dest='verbosity', action='store', default=1, type=int)
-    parser.add_options(NoseTestSuiteRunner.options)
     (options, args) = parser.parse_args()
 
     run_tests(*args, **options.__dict__)
