+++ settings.py
@@ -44,6 +44,8 @@ def configure_settings():
         ]
 
         settings.configure(
+            TEST_RUNNER='django_nose.NoseTestSuiteRunner',
+            NOSE_ARGS=['--nocapture', '--nologcapture', '--verbosity=1'],
             DATABASES={
                 'default': db_config,
             },
@@ -52,5 +54,4 @@ def configure_settings():
             ROOT_URLCONF='entity.urls',
             DEBUG=False,
             DDF_FILL_NULLABLE_FIELDS=False,
-            NOSE_ARGS=['--nocapture', '--nologcapture', '--verbosity=1'],
         )
